sensor: make runtime configurable attrs continuous
Runtime configurable attributes are now matched with the smallest
available value that is greater or equal to the parameter instead of
returning -EINVAL.


Former-commit-id: 676b89cd1affc3ae710f2b812c3ab12460247412

+++ drivers/sensor/sensor_lsm9ds0_gyro.h
@@ -152,7 +152,7 @@ DECL|LSM9DS0_GYRO_VAL_WHO_AM_I_G|macro|LSM9DS0_GYRO_VAL_WHO_AM_I_G
 DECL|__SENSOR_LSM9DS0_GYRO_H__|macro|__SENSOR_LSM9DS0_GYRO_H__
 DECL|dev|member|struct device *dev;
 DECL|fiber_stack|member|char __stack fiber_stack[CONFIG_LSM9DS0_GYRO_FIBER_STACK_SIZE];
-DECL|fs|member|int fs;
+DECL|fs|member|uint8_t fs;
 DECL|gpio_cb|member|struct gpio_callback gpio_cb;
 DECL|gpio_drdy_dev_name|member|char *gpio_drdy_dev_name;
 DECL|gpio_drdy_int_pin|member|uint8_t gpio_drdy_int_pin;
@@ -163,7 +163,7 @@ DECL|i2c_master|member|struct device *i2c_master;
 DECL|i2c_slave_addr|member|uint16_t i2c_slave_addr;
 DECL|lsm9ds0_gyro_config|struct|struct lsm9ds0_gyro_config {
 DECL|lsm9ds0_gyro_data|struct|struct lsm9ds0_gyro_data {
-DECL|sample_fs|member|int sample_fs;
+DECL|sample_fs|member|uint8_t sample_fs;
 DECL|sample_x|member|int sample_x, sample_y, sample_z;
 DECL|sample_y|member|int sample_x, sample_y, sample_z;
 DECL|sample_z|member|int sample_x, sample_y, sample_z;
